app配置及同步调整

+++ src/main/java/com/tipray/service/impl/AppdevServiceImpl.java
@@ -1,30 +1,18 @@
 package com.tipray.service.impl;
 
-import com.tipray.bean.AppCenter;
 import com.tipray.bean.AppDev;
 import com.tipray.bean.AppInfo;
-import com.tipray.bean.AppSync;
-import com.tipray.bean.AppVer;
-import com.tipray.bean.Center;
 import com.tipray.bean.GridPage;
 import com.tipray.bean.Page;
-import com.tipray.bean.ResponseMsg;
 import com.tipray.core.exception.ServiceException;
 import com.tipray.dao.AppdevDao;
-import com.tipray.dao.AppverDao;
 import com.tipray.service.AppdevService;
 import com.tipray.util.EmptyObjectUtil;
-import com.tipray.util.JSONUtil;
-import com.tipray.util.OkHttpUtil;
 import com.tipray.util.StringUtil;
-import okhttp3.FormBody;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -34,23 +22,11 @@ import java.util.List;
 @Transactional
 @Service("appdevService")
 public class AppdevServiceImpl implements AppdevService {
-    private final Logger logger = LoggerFactory.getLogger(AppdevServiceImpl.class);
     @Resource
     private AppdevDao appdevDao;
-    @Resource
-    private AppverDao appverDao;
 
     @Override
     public AppDev addAppdev(AppDev appDev) throws ServiceException {
-        if (appDev == null) {
-            return null;
-        }
-        appdevDao.add(appDev);
-        return appDev;
-    }
-
-    @Override
-    public AppDev addAppdev(AppDev appDev, String centerIds) throws ServiceException {
         if (appDev == null) {
             return null;
         }
@@ -77,115 +53,12 @@ public class AppdevServiceImpl implements AppdevService {
                 appdevDao.add(appDev);
             }
         }
-        List<Long> dbCenterIds = appdevDao.findCenterIdsByUuidAndAppid(uuid, appid);
-        if (!dbCenterIds.isEmpty()) {
-            appdevDao.deleteAppCenter(uuid, appid, dbCenterIds);
-        }
-        String[] centerIdArr = StringUtil.splitStrWithComma(centerIds);
-        List<Long> newCenter = new ArrayList<>();
-        for (String s : centerIdArr) {
-            Long centerId = Long.parseLong(s, 10);
-            newCenter.add(centerId);
-        }
-        if (!newCenter.isEmpty()) {
-            appdevDao.addAppCenter(uuid, appid, newCenter);
-            for (Long centerId : newCenter) {
-                try {
-                    AppInfo appInfo = appdevDao.getCenterWebAddr(centerId);
-                    String url = new StringBuffer().append("https://").append(appInfo.getIp()).append(':')
-                            .append(appInfo.getPort()).append("/api/appdev/add").toString();
-                    FormBody formBody = new FormBody.Builder().add("appdev", JSONUtil.stringify(appDev)).build();
-                    String reply = OkHttpUtil.post(url, formBody);
-                    ResponseMsg responseMsg = JSONUtil.parseToObject(reply, ResponseMsg.class);
-                    if (responseMsg.getId() > 0) {
-                        logger.warn("向用户中心[{}]同步新增APP设备信息失败：{}", centerId, responseMsg.getMsg());
-                    }
-                } catch (Exception e) {
-                    logger.error("向用户中心[" + centerId + "]同步新增APP设备信息异常！", e);
-                }
-            }
-        }
-
         return appDev;
     }
 
-    @Override
-    public AppDev updateAppdev(AppDev appDev, String centerIds) throws ServiceException {
-        if (appDev == null) {
-            return null;
-        }
-        appdevDao.update(appDev);
-        Long appdevId = appDev.getId();
-        List<Long> dbCenterIds = appdevDao.findCenterIdsByAppdevId(appdevId);
-
-        String[] centerIdArr = StringUtil.splitStrWithComma(centerIds);
-        List<Long> newCenter = new ArrayList<>();
-        List<Long> upCenter = new ArrayList<>();
-
-        for (String s : centerIdArr) {
-            Long centerId = Long.parseLong(s, 10);
-            if (dbCenterIds.contains(centerId)) {
-                upCenter.add(centerId);
-                dbCenterIds.remove(centerId);
-            } else {
-                newCenter.add(centerId);
-            }
-        }
-        appDev = appdevDao.getById(appdevId);
-        String uuid = appDev.getUuid();
-        String appid = appDev.getAppid();
-        if (!newCenter.isEmpty()) {
-            appdevDao.addAppCenter(uuid, appid, newCenter);
-            for (Long centerId : newCenter) {
-                try {
-                    AppInfo appInfo = appdevDao.getCenterWebAddr(centerId);
-                    String url = new StringBuffer().append("https://").append(appInfo.getIp()).append(':')
-                            .append(appInfo.getPort()).append("/api/appdev/add").toString();
-                    FormBody formBody = new FormBody.Builder().add("appdev", JSONUtil.stringify(appDev)).build();
-                    String reply = OkHttpUtil.post(url, formBody);
-                    ResponseMsg responseMsg = JSONUtil.parseToObject(reply, ResponseMsg.class);
-                    if (responseMsg.getId() > 0) {
-                        logger.warn("向用户中心[{}]同步新增APP设备信息失败：{}", centerId, responseMsg.getMsg());
-                    }
-                } catch (Exception e) {
-                    logger.error("向用户中心[" + centerId + "]同步新增APP设备信息异常！", e);
-                }
-            }
-        }
-        if (!dbCenterIds.isEmpty()) {
-            appdevDao.deleteAppCenter(uuid, appid, dbCenterIds);
-            for (Long centerId : dbCenterIds) {
-                try {
-                    AppInfo appInfo = appdevDao.getCenterWebAddr(centerId);
-                    String url = new StringBuffer().append("https://").append(appInfo.getIp()).append(':')
-                            .append(appInfo.getPort()).append("/api/appdev/delete").toString();
-                    FormBody formBody = new FormBody.Builder().add("id", appdevId.toString()).build();
-                    String reply = OkHttpUtil.post(url, formBody);
-                    ResponseMsg responseMsg = JSONUtil.parseToObject(reply, ResponseMsg.class);
-                    if (responseMsg.getId() > 0) {
-                        logger.warn("向用户中心[{}]同步删除APP设备信息失败：{}", centerId, responseMsg.getMsg());
-                    }
-                } catch (Exception e) {
-                    logger.error("向用户中心[" + centerId + "]同步删除APP设备信息异常！", e);
-                }
-            }
-        }
-        if (!upCenter.isEmpty()) {
-            for (Long centerId : upCenter) {
-                try {
-                    AppInfo appInfo = appdevDao.getCenterWebAddr(centerId);
-                    String url = new StringBuffer().append("https://").append(appInfo.getIp()).append(':')
-                            .append(appInfo.getPort()).append("/api/appdev/update").toString();
-                    FormBody formBody = new FormBody.Builder().add("appdev", JSONUtil.stringify(appDev)).build();
-                    String reply = OkHttpUtil.post(url, formBody);
-                    ResponseMsg responseMsg = JSONUtil.parseToObject(reply, ResponseMsg.class);
-                    if (responseMsg.getId() > 0) {
-                        logger.warn("向用户中心[{}]同步更新APP设备信息失败：{}", centerId, responseMsg.getMsg());
-                    }
-                } catch (Exception e) {
-                    logger.error("向用户中心[" + centerId + "]同步更新APP设备信息异常！", e);
-                }
-            }
+    public AppDev updateAppdev(AppDev appDev) throws ServiceException {
+        if (appDev != null) {
+            appdevDao.update(appDev);
         }
         return appDev;
     }
@@ -197,25 +70,8 @@ public class AppdevServiceImpl implements AppdevService {
 
     @Override
     public void deleteAppdevById(Long id) throws ServiceException {
-        List<Long> centerIds = appdevDao.findCenterIdsByAppdevId(id);
-        appdevDao.deleteAppCenterByAppdevId(id);
-        appdevDao.delete(id);
-        if (!centerIds.isEmpty()) {
-            for (Long centerId : centerIds) {
-                try {
-                    AppInfo appInfo = appdevDao.getCenterWebAddr(centerId);
-                    String url = new StringBuffer().append("https://").append(appInfo.getIp()).append(':')
-                            .append(appInfo.getPort()).append("/api/appdev/delete").toString();
-                    FormBody formBody = new FormBody.Builder().add("id", id.toString()).build();
-                    String reply = OkHttpUtil.post(url, formBody);
-                    ResponseMsg responseMsg = JSONUtil.parseToObject(reply, ResponseMsg.class);
-                    if (responseMsg.getId() > 0) {
-                        logger.warn("向用户中心[{}]同步删除APP设备信息失败：{}", centerId, responseMsg.getMsg());
-                    }
-                } catch (Exception e) {
-                    logger.error("向用户中心[" + centerId + "]同步删除APP设备信息异常！", e);
-                }
-            }
+        if (id != null) {
+            appdevDao.delete(id);
         }
     }
 
@@ -258,46 +114,6 @@ public class AppdevServiceImpl implements AppdevService {
         long records = countAppdev(appDev);
         List<AppDev> list = findByPage(appDev, page);
         return new GridPage<>(list, records, page.getPageId(), page.getRows(), list.size(), appDev);
-    }
-
-    @Override
-    public List<String> findCenterNamesByAppdevId(Long appdevId) {
-        return null == appdevId ? null : appdevDao.findCenterNamesByAppdevId(appdevId);
-    }
-
-    @Override
-    public List<Center> findBelongCentersByAppdevId(Long appdevId) {
-        return null == appdevId ? null : appdevDao.findBelongCentersByAppdevId(appdevId);
-    }
-
-    @Override
-    public List<AppCenter> findAppCentersByAppdevId(Long appdevId) {
-        return null == appdevId ? null : appdevDao.findAppCentersByAppdevId(appdevId);
-    }
-
-    @Override
-    public List<Long> findCenterIdsByUuidAndAppid(String uuid, String appid) {
-        return appdevDao.findCenterIdsByUuidAndAppid(uuid, appid);
-    }
-
-    @Override
-    public List<Center> findCentersByUuid(String uuid) {
-        return StringUtil.isEmpty(uuid) ? null : appdevDao.findCentersByUuid(uuid);
-    }
-
-    @Override
-    public List<String> findUuidsByCenterId(Long centerId) {
-        return null == centerId ? null : appdevDao.findUuidsByCenterId(centerId);
-    }
-
-    @Override
-    public AppSync findAppSync(Long centerId) {
-        List<AppDev> appDevs = appdevDao.findByCenterId(centerId);
-        List<AppVer> appVers = appverDao.findByCenterId(centerId);
-        AppSync appSync = new AppSync();
-        appSync.setAppdevs(appDevs);
-        appSync.setAppvers(appVers);
-        return appSync;
     }
 
     @Override
